+++ config/zsh/nvr.zsh
@@ -24,15 +24,17 @@ local __current_tmux_session="$(tmux display-message -p '#S')"
 local __current_session_window="$(tmux display-message -p '#I')"
 # Replace slashes on session name to prevent socket creation errors
 __current_tmux_session="${__current_tmux_session//\//-}"
-export NVIM_LISTEN_ADDRESS="/tmp/nvimsocket-${__current_tmux_session}-${__current_session_window}"
+local __listen_address="/tmp/nvimsocket-${__current_tmux_session}-${__current_session_window}"
 
-alias nvim="nvim --listen $NVIM_LISTEN_ADDRESS"
+alias nvim="nvim --listen $__listen_address"
 
-export EDITOR="nvr --servername $NVIM_LISTEN_ADDRESS"
-export VISUAL="nvr --servername $NVIM_LISTEN_ADDRESS"
+export EDITOR="nvr --servername $__listen_address"
+export VISUAL="nvr --servername $__listen_address"
 
 function e() {
-  nvr --servername $NVIM_LISTEN_ADDRESS "$1" && tmux select-pane -l
+  # TODO: select the pane containing neovim instead of last pane
+  # && tmux select-pane -l
+  nvr --servername $__listen_address "$1"
 }
 
 # Elixir Editor Support
@@ -43,9 +45,17 @@ function e() {
 # probably a better way to do that.)
 # open exh source files in neovim in the last pane
 # see https://github.com/rowlandcodes/exhelp#open-source-code-in-an-editor
-export ELIXIR_EDITOR="$EDITOR +'__LINE__' __FILE__ && tmux select-pane -l"
+# --nostart is an arg to nvr - neovim can't start inside an interactive iex
+# shell so it was hanging
+# -s is an argument to nvr - to tell it not to show the warning when the server
+# doesn't exist
+export ELIXIR_EDITOR="$EDITOR --nostart -s +'__LINE__' __FILE__"
 
 # open ecto generated source files in neovim in the last pane
 # see https://hexdocs.pm/ecto/Mix.Tasks.Ecto.Gen.Repo.html
 # and https://hexdocs.pm/ecto_sql/Mix.Tasks.Ecto.Gen.Migration.html
-export ECTO_EDITOR="$EDITOR +'__LINE__' __FILE__ && tmux select-pane -l"
+# --nostart is an arg to nvr - neovim can't start inside an interactive iex
+# shell so it was hanging
+# -s is an argument to nvr - to tell it not to show the warning when the server
+# doesn't exist
+export ECTO_EDITOR="$EDITOR --nostart -s +'__LINE__' __FILE__"